requirements.txt
google-adk>=1.4.2
python-dotenv>=1.0.0
pyyaml>=6.0
